Move and clean up visual files
