--- src/components/SymbolicAICareViz.tsx
@@ -1,15 +1,0 @@
-import React from 'react';
-
-interface SymbolicAICareVizProps {
-  businessFocus?: 'care' | 'industry';
-}
-
-const SymbolicAICareViz: React.FC<SymbolicAICareVizProps> = ({ businessFocus }) => {
-  return (
-    <div className="flex items-center justify-center h-full w-full bg-pink-100 text-pink-800">
-      <p>{businessFocus === 'care' ? 'Care-focused Symbolic AI Visualization' : 'Industry-focused Symbolic AI Visualization'}</p>
-    </div>
-  );
-};
-
-export default SymbolicAICareViz;
