--- src/components/SymbolicAIIndustryViz.tsx
@@ -1,15 +1,0 @@
-import React from 'react';
-
-interface SymbolicAIIndustryVizProps {
-  businessFocus?: 'care' | 'industry';
-}
-
-const SymbolicAIIndustryViz: React.FC<SymbolicAIIndustryVizProps> = ({ businessFocus }) => {
-  return (
-    <div className="flex items-center justify-center h-full w-full bg-yellow-100 text-yellow-800">
-      <p>{businessFocus === 'care' ? 'Care-focused Symbolic AI Visualization' : 'Industry-focused Symbolic AI Visualization'}</p>
-    </div>
-  );
-};
-
-export default SymbolicAIIndustryViz;
