+++ src/components/Visualizations/technical/FeatureExtractionViz.tsx
