+++ src/components/Visualizations/technical/SymbolicAIViz.tsx
